Move www02 and key share structures to the root level
