+++ synedrion/src/cggmp21/protocols.rs
@@ -1,17 +1,10 @@
-mod common;
 pub(crate) mod interactive_signing;
 pub(crate) mod key_gen;
 pub(crate) mod key_init;
 pub(crate) mod key_refresh;
-pub(crate) mod key_resharing;
 pub(crate) mod presigning;
 pub(crate) mod signing;
-mod threshold;
 
-#[cfg(feature = "bench-internals")]
-pub(crate) use common::PresigningData;
-
-pub use common::{KeyShare, KeyShareChange};
 pub use interactive_signing::{
     InteractiveSigningError, InteractiveSigningProof, InteractiveSigningResult,
 };
@@ -20,4 +13,3 @@ pub use key_init::{KeyInitError, KeyInitResult};
 pub use key_refresh::KeyRefreshResult;
 pub use presigning::{PresigningError, PresigningProof, PresigningResult};
 pub use signing::{SigningProof, SigningResult};
-pub use threshold::ThresholdKeyShare;
