+++ synedrion/src/common.rs
