+++ synedrion/src/lib.rs
@@ -25,12 +25,15 @@ cfg_if::cfg_if! {
     }
 }
 
+mod common;
 mod curve;
 mod paillier;
 mod rounds;
 pub mod sessions;
+mod threshold;
 mod tools;
 mod uint;
+mod www02;
 
 // Some re-exports to avoid the need for version-matching
 pub use k256;
@@ -39,9 +42,11 @@ pub use signature;
 
 pub use cggmp21::{
     InteractiveSigningError, InteractiveSigningProof, InteractiveSigningResult, KeyGenError,
-    KeyGenProof, KeyGenResult, KeyInitError, KeyInitResult, KeyRefreshResult, KeyShare,
-    KeyShareChange, PresigningError, PresigningProof, PresigningResult, ProductionParams,
-    SchemeParams, SigningProof, SigningResult, TestParams, ThresholdKeyShare,
+    KeyGenProof, KeyGenResult, KeyInitError, KeyInitResult, KeyRefreshResult, PresigningError,
+    PresigningProof, PresigningResult, ProductionParams, SchemeParams, SigningProof, SigningResult,
+    TestParams,
 };
+pub use common::{KeyShare, KeyShareChange};
 pub use curve::RecoverableSignature;
 pub use rounds::ProtocolResult;
+pub use threshold::ThresholdKeyShare;
